+++ app/resources/Contacts.py
@@ -1,27 +1,41 @@
 from flask import request, jsonify
+from sqlalchemy import exc
 from app import db
 from ..models.Contacts import Contacts, contact_schema, contacts_schema
 from ..models.TypeContacts import TypeContacts
 from ..models.Users import Users
+from ..services.auth import is_your
 
 def post_contact():
     #pegando os campos da requisicao
-    description = request.json['description']
-    typeContact_id = request.json['typeContact_id']
-    user_id = request.json['user_id']
-    typeContact = TypeContacts.query.get(request.json['typeContact_id'])
-    user = Users.query.get(request.json['user_id'])
+    try:
+        contact = request.json['contact']
+        description = request.json['description']
+        type_contact_id = request.json['type_contact_id']
+        user_id = request.json['user_id']
+        typeContact = TypeContacts.query.get(request.json['type_contact_id'])
+        user = Users.query.get(request.json['user_id'])
+    except Exception as e:
+        return jsonify({'message': 'Expected contact, description, type_contact_id and user_id'}), 400
 
-    contact = Contacts(description)
-    contact.type_contacts = typeContact
-    contact.users = user
+    if not user or not typeContact:
+        return jsonify({'message': 'type_contact_id or user_id does not exist'}), 400
+
+    contact = Contacts(contact, description)
+    contact.type_contact = typeContact
+    contact.user = user
     try:
         db.session.add(contact)#adiciona
         db.session.commit()# commit no banco
         result = contact_schema.dump(contact)
         return jsonify({'message': 'Sucessfully registered', 'data': result}), 201
-    except Exception as e:
+    except exc.IntegrityError as e:
+        if 'Duplicate entry' in e.orig.args[1]:#se isso for true, significa que teve duplicida e nesse caso so pode ser o contact
+            return jsonify({'message': 'This contact is already in use', 'data': {}}), 406
+        else:
             return jsonify({'message': 'We had an error processing your data, please try again in a few moments', 'data': {}}), 400
+    except:
+        return jsonify({'message': 'We had an error processing your data, please try again in a few moments', 'data': {}}), 400
 
 def update_contact(id):
     contact = Contacts.query.get(id)#procura o contact pelo id
@@ -29,23 +43,30 @@ def update_contact(id):
     if not contact:#se nao existir o contact
         return jsonify({'message': "Contact don't exist", 'data': {}}), 404
 
-    if 'typeContact_id' in request.json:
-        typeContact = TypeContacts.query.get(request.json['typeContact_id'])
+    if not is_your(contact.user_id):
+        return jsonify({'message': "Unauthorized action."}), 401
+
+    if 'type_contact_id' in request.json:
+        typeContact = TypeContacts.query.get(request.json['type_contact_id'])
 
 
     #substitui ou mantem os campos
     contact.description = request.json['description'] if 'description' in request.json else contact.description
-    contact.type_contacts = typeContacts.query.get(request.json['typeContact_id']) if 'typeContact_id' in request.json else contact.type_contacts
+    contact.type_contact = typeContacts.query.get(request.json['type_contact_id']) if 'type_contact_id' in request.json else contact.type_contact
 
     try:
         db.session.commit()
         result = contact_schema.dump(contact)
         return jsonify({'message': 'Sucessfully updated', 'data': result}), 201
-    except Exception as e:
-        if ('Duplicate entry' in e.orig.args[1]):#se isso for true, significa que teve duplicida e nesse caso so pode ser o name
-            return jsonify({'message': 'This name is already in use', 'data': {}}), 406
+    except exc.IntegrityError as e:
+        if 'Duplicate entry' in e.orig.args[1]:#se isso for true, significa que teve duplicida e nesse caso so pode ser o contact
+            return jsonify({'message': 'This contact is already in use', 'data': {}}), 406
         else:
             return jsonify({'message': 'We had an error processing your data, please try again in a few moments', 'data': {}}), 400
+    except:
+        return jsonify({'message': 'We had an error processing your data, please try again in a few moments', 'data': {}}), 400
+
+
 
 def get_contacts():
     contacts = Contacts.query.all()#pega todos contacts
@@ -59,6 +80,9 @@ def get_contact(id):
     contact = Contacts.query.get(id)#busca contact pelo id
 
     if contact:#se existir
+        print(contact.user_id)
+        if not is_your(contact.user_id):#verifica se o contato e do usuario logado
+            return jsonify({'message': "Unauthorized action."}), 401
         result = contact_schema.dump(contact)
         return jsonify({"message": "Sucessfully fetched", "data": result})
     #se nao existir
@@ -69,6 +93,9 @@ def delete_contact(id):
 
     if not contact:#se nao existir
         return jsonify({'message': "Contact don't exist", 'data': {}}), 404
+
+    if not is_your(contact.user_id):#verifica se o contato e do usuario logado
+        return jsonify({'message': "Unauthorized action."}), 401
 
     try:
         db.session.delete(contact)
